Fixed issued with circular import related to server_defaults.
